FastDifferentiation.clear_cache — Method

Clears the global expression cache. To maximize efficiency of expressions the differentation system automatically eliminates common subexpressions by checking for their existence in the global expression cache. Over time this cache can become arbitrarily large. Best practice is to clear the cache before you start defining expressions, define your expressions and then clear the cache.

FastDifferentiation.derivative — Method

computes ∂A/(∂variables[1],...,∂variables[n]). Repeated differentiation rather than computing different columns of the Jacobian. Example:

julia> A = [t t^2;3t^2 5]  
2×2 Matrix{Node}:
 t              (t ^ 2)
 (3 * (t ^ 2))  5

julia> derivative(A,t)  
2×2 Matrix{Node}:
 1.0      (2 * t)
 (6 * t)  0.0

julia> derivative(A,t,t)  
2×2 Matrix{Node{T, 0} where T}:
 0.0  2
 6    0.0

FastDifferentiation.differential — Method

Returns an anonymous function that takes the derivative of a scalar function with respect to variables. Example:

julia> @variables t
t

julia> f = t^2
(t ^ 2)

julia> Dt = differential(t)      
#69 (generic function with 1 method)

julia> Dt(f)
(2 * t)

julia> Dt = differential(t,t)    
#69 (generic function with 1 method)

julia> Dt(f)
2

FastDifferentiation.hessian — Method

Returns the dense symbolic Hessian matrix. Example:

julia> @variables x y

julia> hessian(x^2*y^2,[x,y])
2×2 Matrix{FastDifferentiation.Node}:
 (2 * (y ^ 2))  (4 * (y * x))
 (4 * (x * y))  (2 * (x ^ 2))

FastDifferentiation.hessian_times_v — Method

Computes Hessian times a vector v without forming the Hessian matrix. Useful when the Hessian would be impractically large.

FastDifferentiation.jacobian — Method

Jacobian matrix of the n element function defined by terms. Each term element is a Node expression graph. Only the columns of the Jacobian corresponsing to the elements of partial_variables will be computed and the partial columns in the Jacobian matrix will be in the order specified by partial_variables. Examples:

julia> @variables x y

julia> jacobian([x*y,y*x],[x,y])
2×2 Matrix{Node}:
 y  x
 y  x

julia> jacobian([x*y,y*x],[y,x])
2×2 Matrix{Node}:
 x  y
 x  y

julia> jacobian([x*y,y*x],[x])
2×1 Matrix{Node}:
 y
 y

FastDifferentiation.jacobian_times_v — Method

Returns a vector of Node, where each element in the vector is the symbolic form of Jv. Also returnsv_vectora vector of thevvariables. This is useful if you want to generate a function to evaluateJvand you want to separate the inputs to the function and thev` variables.

FastDifferentiation.jacobian_transpose_v — Method

Returns a vector of Node, where each element in the vector is the symbolic form of Jᵀv. Also returns v_vector a vector of the v variables. This is useful if you want to generate a function to evaluate Jᵀv and you want to separate the inputs to the function and the v variables.

FastDifferentiation.make_Expr — Method

init_with_zeros argument is not used for sparse matrices.

FastDifferentiation.make_function — Method

make_function(func_array::AbstractArray{T}, input_variables::AbstractVector{<:Node}...; in_place::Bool=false, init_with_zeros::Bool=true) where {T<:Node}

Makes a function to evaluate the symbolic expressions in func_array. Every variable that is used in func_array must also be in input_variables. However, it will not cause an error if variables ininputvariablesare not variables used byfuncarray`.

If `in_place=false` then the returned array will be properly initialized with zeros.

If in_place=true it will generate code to fill a user supplied array with the result.

If the array is dense and in_place=true then the keyword argument init_with_zeros affects how the in place array is initialized. If init_with_zeros = true then the in place array is initialized with zeros. If init_with_zeros=false it is the user's responsibility to initialize the array with zeros before passing it to the runtime generated function.

This can be useful for modestly sparse dense matrices with say at least 1/4 of the array entries non-zero. In this case a sparse matrix may not be as efficient as a dense matrix. But a large fraction of time could be spent unnecessarily setting elements to zero. In this case you can initialize the in place Jacobian array once with zeros before calling the run time generated function.

julia> @variables x
x

julia> f = x+1
(x + 1)

julia> jac = jacobian([f],[x])
1×1 Matrix{FastDifferentiation.Node}:
 1

julia> jac #the Jacobian has a single constant element, 1, and is no longer a function of x
1×1 Matrix{FastDifferentiation.Node}:
 1

julia> make_function(jac,[x]) #But you can specify x as an input variable to the runtime generated function.

FastDifferentiation.make_variables — Method

Returns an Array of variables with names corresponding to their indices in the Array. Example: ```julia julia> make_variables(:x,3) 3-element Vector{FastDifferentiation.Node}: x1 x2 x3

julia> makevariables(:x,2,3) 2×3 Matrix{FastDifferentiation.Node}: x11 x12 x13 x21 x22 x2_3

julia> makevariables(:x,2,3,2) 2×3×2 Array{FastDifferentiation.Node, 3}: [:, :, 1] = x111 x121 x131 x211 x221 x23_1

[:, :, 2] = x112 x122 x132 x212 x222 x232 ```

FastDifferentiation.sparse_hessian — Method

Compute a sparse symbolic Hessian. Returns a sparse matrix of symbolic expressions. Can be used in combination with make_function to generate an executable that will return a sparse matrix or take one as an in-place argument. Example:

julia> @variables x y

julia> a = sparse_hessian(x*y,[x,y])
2×2 SparseArrays.SparseMatrixCSC{FastDifferentiation.Node, Int64} with 2 stored entries:
 ⋅  1
 1  ⋅

julia> f1 = make_function(a,[x,y])
...

julia> f1([1.0,2.0])
2×2 SparseArrays.SparseMatrixCSC{Float64, Int64} with 2 stored entries:
  ⋅   1.0
 1.0   ⋅

julia> tmp = similar(a,Float64)
2×2 SparseArrays.SparseMatrixCSC{Float64, Int64} with 2 stored entries:
  ⋅            4.24399e-314
 4.24399e-314   ⋅

julia> f2 = make_function(a,[x,y],in_place=true)
...

julia> f2([1.0,2.0],tmp)
2×2 SparseArrays.SparseMatrixCSC{Float64, Int64} with 2 stored entries:
  ⋅   1.0
 1.0   ⋅

julia> tmp
2×2 SparseArrays.SparseMatrixCSC{Float64, Int64} with 2 stored entries:
  ⋅   1.0
 1.0   ⋅

FastDifferentiation.sparse_jacobian — Method

Returns a sparse array containing the Jacobian of the function defined by terms

FastDifferentiation.sparsity — Method

Computes a number representing the sparsity of the array of expressions. If nelts is the number of elements in the array and nzeros is the number of zero elements in the array then sparsity = (nelts-nzeros)/nelts.

Frequently used in combination with a call to make_function to determine whether to set keyword argument init_with_zeros to false.