FastDifferentiation.clear_cache — Method

Clears the global expression cache. To maximize efficiency of expressions the differentation system automatically eliminates common subexpressions by checking for there existence in the global expression cache. Over time this cache can become arbitrarily large. Best practice is to clear the cache before you start defining expressions, define your expressions and then clear the cache.

FastDifferentiation.derivative — Method

computes ∂A/∂variables[1],...,variables[n]. Repeated differentiation rather than computing different columns of the Jacobian. Example:

```julia_repl

julia> A = [t t^2;3t^2 5] 2×2 Matrix{Node}: t (t ^ 2) (3 * (t ^ 2)) 5

julia> derivative(A,t) 2×2 Matrix{Node}: 1.0 (2 * t) (6 * t) 0.0

julia> derivative(A,t,t) 2×2 Matrix{Node{T, 0} where T}: 0.0 2 6 0.0 ```

FastDifferentiation.hessian — Method

Computes the full symbolic Hessian matrix

FastDifferentiation.hessian — Method

Computes the full symbolic Hessian matrix

FastDifferentiation.hessian_times_v — Method

Computes Hessian times a vector v without forming the Hessian matrix. Useful when the Hessian would be impractically large.

FastDifferentiation.jacobian — Method

Jacobian matrix of the n element function defined by terms. Each term element is a Node expression graph. Only the columns of the Jacobian corresponsing to the elements of partial_variables will be computed and the partial columns in the Jacobian matrix will be in the order specified by partial_variables. Examples:

julia> @variables x y

julia> jacobian([x*y,y*x],[x,y])
2×2 Matrix{Node}:
 y  x
 y  x

julia> jacobian([x*y,y*x],[y,x])
2×2 Matrix{Node}:
 x  y
 x  y

julia> jacobian([x*y,y*x],[x])
2×1 Matrix{Node}:
 y
 y

FastDifferentiation.jacobian_times_v — Method

Returns a vector of Node, where each element in the vector is the symbolic form of Jv. Also returnsv_vectora vector of thevvariables. This is useful if you want to generate a function to evaluateJvand you want to separate the inputs to the function and thev` variables.

FastDifferentiation.jacobian_transpose_v — Method

Returns a vector of Node, where each element in the vector is the symbolic form of Jᵀv. Also returns v_vector a vector of the v variables. This is useful if you want to generate a function to evaluate Jᵀv and you want to separate the inputs to the function and the v variables.

FastDifferentiation.sparse_jacobian — Method

Returns a sparse array containing the Jacobian of the function defined by terms

FastDifferentiation.sparsity — Method

Computes sparsity of Jacobian matrix = nonzeroentries/total_entries.